Move course search to shared function
